distributing classes into their own files
