+++ broadlink/__init__.py
@@ -12,7 +12,14 @@ import socket
 import sys
 import threading
 import codecs
-from device import device
+
+from device     import device
+from mp_dev     import mp1
+from sp_dev     import sp1,sp2
+from dooya_dev  import dooya
+from s1_dev     import S1C
+from rm_dev     import rm
+from a_dev      import a1
 
 def gendevice(devtype, host, mac):
   if devtype == 0: # SP1
@@ -140,347 +147,6 @@ def discover(timeout=None, local_ip_address=None):
       dev = gendevice(devtype, host, mac)
       devices.append(dev)
     return devices
-
-
-
-
-class mp1(device):
-  def __init__ (self, host, mac):
-    device.__init__(self, host, mac)
-    self.type = "MP1"
-
-  def set_power_mask(self, sid_mask, state):
-    """Sets the power state of the smart power strip."""
-
-    packet = bytearray(16)
-    packet[0x00] = 0x0d
-    packet[0x02] = 0xa5
-    packet[0x03] = 0xa5
-    packet[0x04] = 0x5a
-    packet[0x05] = 0x5a
-    packet[0x06] = 0xb2 + ((sid_mask<<1) if state else sid_mask)
-    packet[0x07] = 0xc0
-    packet[0x08] = 0x02
-    packet[0x0a] = 0x03
-    packet[0x0d] = sid_mask
-    packet[0x0e] = sid_mask if state else 0
-
-    response = self.send_packet(0x6a, packet)
-
-    err = response[0x22] | (response[0x23] << 8)
-
-  def set_power(self, sid, state):
-    """Sets the power state of the smart power strip."""
-    sid_mask = 0x01 << (sid - 1)
-    return self.set_power_mask(sid_mask, state)
-
-  def check_power_raw(self):
-    """Returns the power state of the smart power strip in raw format."""
-    packet = bytearray(16)
-    packet[0x00] = 0x0a
-    packet[0x02] = 0xa5
-    packet[0x03] = 0xa5
-    packet[0x04] = 0x5a
-    packet[0x05] = 0x5a
-    packet[0x06] = 0xae
-    packet[0x07] = 0xc0
-    packet[0x08] = 0x01
-
-    response = self.send_packet(0x6a, packet)
-    err = response[0x22] | (response[0x23] << 8)
-    if err == 0:
-      payload = self.decrypt(bytes(response[0x38:]))
-      if type(payload[0x4]) == int:
-        state = payload[0x0e]
-      else:
-        state = ord(payload[0x0e])
-      return state
-
-  def check_power(self):
-    """Returns the power state of the smart power strip."""
-    state = self.check_power_raw()
-    data = {}
-    data['s1'] = bool(state & 0x01)
-    data['s2'] = bool(state & 0x02)
-    data['s3'] = bool(state & 0x04)
-    data['s4'] = bool(state & 0x08)
-    return data
-
-
-class sp1(device):
-  def __init__ (self, host, mac):
-    device.__init__(self, host, mac)
-    self.type = "SP1"
-
-  def set_power(self, state):
-    packet = bytearray(4)
-    packet[0] = state
-    self.send_packet(0x66, packet)
-
-
-class sp2(device):
-  def __init__ (self, host, mac):
-    device.__init__(self, host, mac)
-    self.type = "SP2"
-
-  def set_power(self, state):
-    """Sets the power state of the smart plug."""
-    packet = bytearray(16)
-    packet[0] = 2
-    packet[4] = 1 if state else 0
-    self.send_packet(0x6a, packet)
-
-  def check_power(self):
-    """Returns the power state of the smart plug."""
-    packet = bytearray(16)
-    packet[0] = 1
-    response = self.send_packet(0x6a, packet)
-    err = response[0x22] | (response[0x23] << 8)
-    if err == 0:
-      payload = self.decrypt(bytes(response[0x38:]))
-      if type(payload[0x4]) == int:
-        state = bool(payload[0x4])
-      else:
-        state = bool(ord(payload[0x4]))
-      return state
-
-  def get_energy(self):
-    packet = bytearray([8, 0, 254, 1, 5, 1, 0, 0, 0, 45])
-    response = self.send_packet(0x6a, packet)
-    err = response[0x22] | (response[0x23] << 8)
-    if err == 0:
-      payload = self.decrypt(bytes(response[0x38:]))
-      energy = int(hex(ord(payload[7]) * 256 + ord(payload[6]))[2:]) + int(hex(ord(payload[5]))[2:])/100.0
-      return energy
-
-
-class a1(device):
-  def __init__ (self, host, mac):
-    device.__init__(self, host, mac)
-    self.type = "A1"
-
-  def check_sensors(self):
-    packet = bytearray(16)
-    packet[0] = 1
-    response = self.send_packet(0x6a, packet)
-    err = response[0x22] | (response[0x23] << 8)
-    if err == 0:
-      data = {}
-      payload = self.decrypt(bytes(response[0x38:]))
-      if type(payload[0x4]) == int:
-        data['temperature'] = (payload[0x4] * 10 + payload[0x5]) / 10.0
-        data['humidity'] = (payload[0x6] * 10 + payload[0x7]) / 10.0
-        light = payload[0x8]
-        air_quality = payload[0x0a]
-        noise = payload[0xc]
-      else:
-        data['temperature'] = (ord(payload[0x4]) * 10 + ord(payload[0x5])) / 10.0
-        data['humidity'] = (ord(payload[0x6]) * 10 + ord(payload[0x7])) / 10.0
-        light = ord(payload[0x8])
-        air_quality = ord(payload[0x0a])
-        noise = ord(payload[0xc])
-      if light == 0:
-        data['light'] = 'dark'
-      elif light == 1:
-        data['light'] = 'dim'
-      elif light == 2:
-        data['light'] = 'normal'
-      elif light == 3:
-        data['light'] = 'bright'
-      else:
-        data['light'] = 'unknown'
-      if air_quality == 0:
-        data['air_quality'] = 'excellent'
-      elif air_quality == 1:
-        data['air_quality'] = 'good'
-      elif air_quality == 2:
-        data['air_quality'] = 'normal'
-      elif air_quality == 3:
-        data['air_quality'] = 'bad'
-      else:
-        data['air_quality'] = 'unknown'
-      if noise == 0:
-        data['noise'] = 'quiet'
-      elif noise == 1:
-        data['noise'] = 'normal'
-      elif noise == 2:
-        data['noise'] = 'noisy'
-      else:
-        data['noise'] = 'unknown'
-      return data
-
-  def check_sensors_raw(self):
-    packet = bytearray(16)
-    packet[0] = 1
-    response = self.send_packet(0x6a, packet)
-    err = response[0x22] | (response[0x23] << 8)
-    if err == 0:
-      data = {}
-      payload = self.decrypt(bytes(response[0x38:]))
-      if type(payload[0x4]) == int:
-        data['temperature'] = (payload[0x4] * 10 + payload[0x5]) / 10.0
-        data['humidity'] = (payload[0x6] * 10 + payload[0x7]) / 10.0
-        data['light'] = payload[0x8]
-        data['air_quality'] = payload[0x0a]
-        data['noise'] = payload[0xc]
-      else:
-        data['temperature'] = (ord(payload[0x4]) * 10 + ord(payload[0x5])) / 10.0
-        data['humidity'] = (ord(payload[0x6]) * 10 + ord(payload[0x7])) / 10.0
-        data['light'] = ord(payload[0x8])
-        data['air_quality'] = ord(payload[0x0a])
-        data['noise'] = ord(payload[0xc])
-      return data
-
-
-class rm(device):
-  def __init__ (self, host, mac):
-    device.__init__(self, host, mac)
-    self.type = "RM2"
-
-  def check_data(self):
-    packet = bytearray(16)
-    packet[0] = 4
-    response = self.send_packet(0x6a, packet)
-    err = response[0x22] | (response[0x23] << 8)
-    if err == 0:
-      payload = self.decrypt(bytes(response[0x38:]))
-      return payload[0x04:]
-
-  def send_data(self, data):
-    packet = bytearray([0x02, 0x00, 0x00, 0x00])
-    packet += data
-    self.send_packet(0x6a, packet)
-
-  def enter_learning(self):
-    packet = bytearray(16)
-    packet[0] = 3
-    self.send_packet(0x6a, packet)
-
-  def check_temperature(self):
-    packet = bytearray(16)
-    packet[0] = 1
-    response = self.send_packet(0x6a, packet)
-    err = response[0x22] | (response[0x23] << 8)
-    if err == 0:
-      payload = self.decrypt(bytes(response[0x38:]))
-      if type(payload[0x4]) == int:
-        temp = (payload[0x4] * 10 + payload[0x5]) / 10.0
-      else:
-        temp = (ord(payload[0x4]) * 10 + ord(payload[0x5])) / 10.0
-      return temp
-
-# For legay compatibility - don't use this
-class rm2(rm):
-  def __init__ (self):
-    device.__init__(self, None, None)
-
-  def discover(self):
-    dev = discover()
-    self.host = dev.host
-    self.mac = dev.mac
-
-
-S1C_SENSORS_TYPES = {
-    0x31: 'Door Sensor',  # 49 as hex
-    0x91: 'Key Fob',  # 145 as hex, as serial on fob corpse
-    0x21: 'Motion Sensor'  # 33 as hex
-}
-
-
-class S1C(device):
-  """
-  Its VERY VERY VERY DIRTY IMPLEMENTATION of S1C
-  """
-  def __init__(self, *a, **kw):
-    device.__init__(self, *a, **kw)
-    self.type = 'S1C'
-
-  def get_sensors_status(self):
-    packet = bytearray(16)
-    packet[0] = 0x06  # 0x06 - get sensors info, 0x07 - probably add sensors
-    response = self.send_packet(0x6a, packet)
-    err = response[0x22] | (response[0x23] << 8)
-    if err == 0:
-      aes = AES.new(bytes(self.key), AES.MODE_CBC, bytes(self.iv))
-
-      payload = aes.decrypt(bytes(response[0x38:]))
-      if payload:
-        head = payload[:4]
-        count = payload[0x4] #need to fix for python 2.x
-        sensors = payload[0x6:]
-        sensors_a = [bytearray(sensors[i * 83:(i + 1) * 83]) for i in range(len(sensors) // 83)]
-
-        sens_res = []
-        for sens in sensors_a:
-          status = ord(chr(sens[0]))
-          _name = str(bytes(sens[4:26]).decode())
-          _order = ord(chr(sens[1]))
-          _type = ord(chr(sens[3]))
-          _serial = bytes(codecs.encode(sens[26:30],"hex")).decode()
-
-          type_str = S1C_SENSORS_TYPES.get(_type, 'Unknown')
-
-          r = {
-            'status': status,
-            'name': _name.strip('\x00'),
-            'type': type_str,
-            'order': _order,
-            'serial': _serial,
-          }
-          if r['serial'] != '00000000':
-            sens_res.append(r)
-        result = {
-          'count': count,
-          'sensors': sens_res
-        }
-        return result
-
-
-class dooya(device):
-  def __init__ (self, host, mac):
-    device.__init__(self, host, mac)
-    self.type = "Dooya DT360E"
-
-  def _send(self, magic1, magic2):
-    packet = bytearray(16)
-    packet[0] = 0x09
-    packet[2] = 0xbb
-    packet[3] = magic1
-    packet[4] = magic2
-    packet[9] = 0xfa
-    packet[10] = 0x44
-    response = self.send_packet(0x6a, packet)
-    err = response[0x22] | (response[0x23] << 8)
-    if err == 0:
-      payload = self.decrypt(bytes(response[0x38:]))
-      return ord(payload[4])
-
-  def open(self):
-    return self._send(0x01, 0x00)
-
-  def close(self):
-    return self._send(0x02, 0x00)
-
-  def stop(self):
-    return self._send(0x03, 0x00)
-
-  def get_percentage(self):
-    return self._send(0x06, 0x5d)
-
-  def set_percentage_and_wait(self, new_percentage):
-    current = self.get_percentage()
-    if current > new_percentage:
-      self.close()
-      while current is not None and current > new_percentage:
-        time.sleep(0.2)
-        current = self.get_percentage()
-
-    elif current < new_percentage:
-      self.open()
-      while current is not None and current < new_percentage:
-        time.sleep(0.2)
-        current = self.get_percentage()
-    self.stop()
 
 # Setup a new Broadlink device via AP Mode. Review the README to see how to enter AP Mode.
 # Only tested with Broadlink RM3 Mini (Blackbean)
